Add I2C driver and sensor implementations for ADXL345, BMP280, GY271, and ICM20948
- Implemented I2C class for I2C communication with configurable parameters.
- Added ADXL345 sensor driver with methods for initialization, reading raw data, and acceleration.
- Created BMP280 sensor driver for temperature and pressure readings with calibration handling.
- Developed GY271 magnetometer driver for reading magnetic field data.
- Implemented ICM20948 sensor driver for accelerometer and gyroscope data retrieval.
- Each sensor driver includes methods for initialization, data reading, and connection verification.

+++ Drivers/Interfaces/IPwm.h
@@ -18,5 +18,5 @@ protected:
 public:
     virtual ~IPwm() = default;
     virtual bool initialize() = 0;
-    virtual uint64_t setPulseWidth(uint32_t pulseWidthUs) = 0;
+    virtual uint32_t setPulseWidth(uint32_t pulseWidthUs) = 0;
 };
